Implemented read from db and write to activity

+++ app/src/main/java/com/example/persistence/Country.java
@@ -2,19 +2,19 @@ package com.example.persistence;
 
 public class Country {
 
-    private final int id;
+    private final long id;
     private String name;
     private String capital;
     private int population;
 
-    public Country(int id, String name, String capital, int population) {
+    public Country(long id, String name, String capital, int population) {
         this.id = id;
         this.name = name;
         this.capital = capital;
         this.population = population;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
